Fix parsing API parameters without type
If there is no specific type specified for an @apiParam, the parsing
fails because it tries to find '=' within the type variable and if it is
None, it can't iterate to find the equals sign.

So whenever the type_ group from the regex now is None, we just set the
parsed type to None and also make sure we handle this properly when
rendering the docstring.

In more recent versions of Habitica the following @apiParam was
introduced, which exactly triggers this problem:

@apiParam (Query) [dueDate] type Optional date to use for computing the
                                 nextDue field for each returned task.

Signed-off-by: aszlig <aszlig@nix.build>